requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.10.0
openai>=1.50.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
requests>=2.31.0
google-generativeai>=0.8.0
